remove model.py because we have the directory
