+++ src/backend/models/user.py
@@ -23,6 +23,30 @@ class User(db.Model):
     a_user = db.relationship("Auctions", back_populates="a_user")
     b_user = db.relationship("Bids", back_populates="b_user")
 
+@dataclass
+class UserQuestion(db.Model):
+    __tablename__ = 'user_questions'
+    id = db.Column(db.Integer, primary_key=True)
+    auction_id = db.Column(db.Integer, db.ForeignKey('auctions.id'), nullable=False)
+    asker_id = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=False)
+    replier_id = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=True)
+    question_text = db.Column(db.String(255), nullable=False)
+    is_open = db.Column(db.Boolean, nullable=False)
+
+    def __repr__(self):
+        return f"<UserQuestion {self.id}, Auction {self.auction_id}, Asker {self.asker_id}>"
+
+@dataclass
+class UserAnswer(db.Model):
+    __tablename__ = 'user_answers'
+    id = db.Column(db.Integer, primary_key=True)
+    question_id = db.Column(db.Integer, db.ForeignKey('user_questions.id'), nullable=False)
+    replier_id = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=False)
+    reply_text = db.Column(db.String(255), nullable=False)
+
+    def __repr__(self):
+        return f"<UserAnswer {self.id}, Question {self.question_id}, Replier {self.replier_id}>"
+
 def add_new(email, username, password, role=constants.USER_ROLE.USER):
     user = User.query.filter((User.email==email) | (User.username==username)).first()
     if (user is None) and email and username and password:
@@ -32,3 +56,5 @@ def add_new(email, username, password, role=constants.USER_ROLE.USER):
         return True
 
     return False
+
+
